# src/docs-openapi.yaml
openapi: 3.0.3
info:
  title: SMKN4 Profile API
  version: 0.1.1
  description: REST API for SMKN4 profile content management.
servers:
  - url: http://api.smkn4.sch.id
    description: Main API
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
...
    FacilityStatus:
      type: string
      enum: [TERSEDIA, PERBAIKAN, DIGUNAKAN, TIDAK_TERSEDIA]
    Facility:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        location: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/FacilityStatus" }
        image_url: { type: string, nullable: true }
    Teacher:
      type: object
      properties:
        guru_id: { type: string, format: uuid }
        name: { type: string }
        jabatan: { type: string }
        nip: { type: string, nullable: true }
        image_url: { type: string, nullable: true }
...
    Slide:
      type: object
      properties:
        id: { type: string, format: uuid }
        image: { type: string }
        alt: { type: string }
        title: { type: string }
        subtitle: { type: string }
        description: { type: string }
        gradientFrom: { type: string }
        gradientTo: { type: string }
        order: { type: integer }
        isActive: { type: boolean }
...
paths:
...
  /teachers/:
    get:
      summary: List/search teachers
      parameters:
        - in: query
          name: nama
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200": { description: OK }
    post:
      summary: Create teacher
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, jabatan]
              properties:
                name: { type: string }
                jabatan: { type: string }
                nip: { type: string }
                image: { type: string, format: binary }
      responses:
        "200":
          {
            description: Created,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Teacher" } },
              },
          }
  /teachers/{id}:
    put:
      summary: Update teacher
      security: [{ cookieAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name: { type: string }
                jabatan: { type: string }
                image: { type: string, format: binary }
      responses:
        "200": { description: Updated }
    delete:
      summary: Delete teacher
      security: [{ cookieAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Deleted }
  /facilities/:
    get:
      summary: List/search facilities
      parameters:
        - in: query
          name: name
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200": { description: OK }
    post:
      summary: Create facility
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, status]
              properties:
                name: { type: string }
                description: { type: string }
                location: { type: string }
                status: { $ref: "#/components/schemas/FacilityStatus" }
                image: { type: string, format: binary }
      responses:
        "200":
          {
            description: Created,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Facility" } },
              },
          }
  /facilities/{id}:
    put:
      summary: Update facility
      security: [{ cookieAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                location: { type: string }
                status: { $ref: "#/components/schemas/FacilityStatus" }
                image: { type: string, format: binary }
      responses:
        "200": { description: Updated }
    delete:
      summary: Delete facility
      security: [{ cookieAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Deleted }
...
  /slides/:
    get:
      summary: List all slides ordered by order field
      responses:
        "200":
          description: Array of slides
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Slide" }
    post:
      summary: Create slide
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                [
                  alt,
                  title,
                  subtitle,
                  description,
                  gradientFrom,
                  gradientTo,
                  order,
                ]
              properties:
                image: { type: string, format: binary }
                alt: { type: string }
                title: { type: string }
                subtitle: { type: string }
                description: { type: string }
                gradientFrom: { type: string }
                gradientTo: { type: string }
                order: { type: integer }
                isActive: { type: boolean }
      responses:
        "200":
          {
            description: Created,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Slide" } },
              },
          }
  /slides/{id}:
    put:
      summary: Update slide
      security: [{ cookieAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [alt, title]
              properties:
                image: { type: string, format: binary }
                alt: { type: string }
                title: { type: string }
                subtitle: { type: string }
                description: { type: string }
                gradientFrom: { type: string }
                gradientTo: { type: string }
                order: { type: integer }
                isActive: { type: boolean }
      responses:
        "200": { description: Updated }
    delete:
      summary: Delete slide
      security: [{ cookieAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Deleted }
...
security: []